Move fixtures to top level in acceptance tests
The fixtures were buried witin a setup folder within the features
folder.

This change moves them to the top level as they are something shared
between step definitions and page objects.
